fix timezone and recurring event issues

+++ server-v2/PortalCalendarServer/Controllers/UiController.cs
@@ -95,9 +95,11 @@ public class UiController(
             return NotFound();
         }
 
-        var configs = _context.Configs.Where(c => c.DisplayId == display.Id);
-        _context.Configs.RemoveRange(configs);
-        _context.Themes.Remove(display.Theme);
+        _context.Configs.RemoveRange(_context.Configs.Where(c => c.DisplayId == display.Id));
+        if (display.Theme != null)
+        {
+            _context.Themes.Remove(display.Theme);
+        }
         _context.Displays.Remove(display);
         await _context.SaveChangesAsync();
